+++ samples/com/landawn/abacus/samples/PerformanceTestByProfiler.java
@@ -1,4 +1,4 @@
-package com.landawn.samples.abacus;
+package com.landawn.abacus.samples;
 
 import org.junit.Test;